add followers relationship

+++ app/helpers/user_helper.rb
@@ -1,5 +1,6 @@
 module UserHelper
     def follow_button(user)
+        return if user.id == current_user.id
         if current_user.following?(user)
             button_to "Unfollow", unfollow_user_path, method: :delete
         else
